+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add getProto method to Inworld packets
 - Support SPEECH_RECOGNITION_ONLY mode
 - Add version to session state
+- Allow to get session state as a Uint8Array
 
 ## [1.14.2] - 2024-06-27